Rename the slsa package to amber
