+++ amber/BUILD
@@ -19,12 +19,12 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 package(default_visibility = ["//:__subpackages__"])
 
 go_library(
-    name = "slsa",
-    srcs = ["slsa.go"],
+    name = "amber",
+    srcs = ["amber.go"],
     data = [
         "//schema/amber-slsa-buildtype/v1:provenance.json",
     ],
-    importpath = "github.com/project-oak/transparent-release/slsa",
+    importpath = "github.com/project-oak/transparent-release/amber",
     deps = [
         "@com_github_in_toto_in_toto_golang//in_toto:go_default_library",
         "@com_github_in_toto_in_toto_golang//in_toto/slsa_provenance/v0.2:go_default_library",
@@ -33,14 +33,14 @@ go_library(
 )
 
 go_test(
-    name = "slsa_test",
+    name = "amber_test",
     size = "small",
-    srcs = ["slsa_test.go"],
+    srcs = ["amber_test.go"],
     data = [
         "//schema/amber-slsa-buildtype/v1:example.json",
         "//schema/amber-slsa-buildtype/v1:provenance.json",
     ],
-    embed = [":slsa"],
+    embed = [":amber"],
     deps = [
         "@com_github_google_go_cmp//cmp:go_default_library",
         "@com_github_in_toto_in_toto_golang//in_toto/slsa_provenance/v0.2:go_default_library",
