+++ common/common_test.go
@@ -22,7 +22,7 @@ import (
 
 	cmp "github.com/google/go-cmp/cmp"
 	slsa "github.com/in-toto/in-toto-golang/in_toto/slsa_provenance/v0.2"
-	amber "github.com/project-oak/transparent-release/slsa"
+	"github.com/project-oak/transparent-release/amber"
 )
 
 const testdataPath = "../testdata/"
