+++ experimental/auth-logic/wrappers/BUILD
@@ -46,7 +46,7 @@ go_library(
     importpath = "github.com/project-oak/transparent-release/experimental/auth-logic/wrappers",
     deps = [
         "//common",
-        "//slsa",
+        "//amber",
         "//verify",
         "@com_github_cyberphone_json_canonicalization//go/src/webpki.org/jsoncanonicalizer",
         "@com_github_go_openapi_runtime//:go_default_library",
