+++ verify/BUILD
@@ -24,7 +24,7 @@ go_library(
     importpath = "github.com/project-oak/transparent-release/verify",
     deps = [
         "//common",
-        "//slsa",
+        "//amber",
         "@com_github_in_toto_in_toto_golang//in_toto/slsa_provenance/v0.2:go_default_library",
     ],
 )
