+++ verify/verify.go
@@ -19,8 +19,8 @@ import (
 	"fmt"
 
 	slsa "github.com/in-toto/in-toto-golang/in_toto/slsa_provenance/v0.2"
+	"github.com/project-oak/transparent-release/amber"
 	"github.com/project-oak/transparent-release/common"
-	amber "github.com/project-oak/transparent-release/slsa"
 )
 
 // ProvenanceVerifier defines an interface with a single method `Verify` for
